giga zmiany

+++ src/Services/CoreService/Repositories/Classes/UserRepository.cs
@@ -1,7 +1,5 @@
-﻿using AutoMapper;
-using ChatLab.CoreService.DbManager;
+﻿using ChatLab.CoreService.DbManager;
 using ChatLab.CoreService.Entities;
-using ChatLab.CoreService.Models.DTOs;
 using ChatLab.CoreService.Repositories.Interfaces;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
@@ -12,49 +10,43 @@ namespace ChatLab.CoreService.Repositories.Classes
     public class UserRepository : IUserRepository
     {
         private readonly AppDbContext _dbContext;
-        private readonly IMapper _mapper;
         private readonly UserManager<User> _userManager;
         private readonly IPasswordHasher<User> _passwordHasher;
 
-        public UserRepository(AppDbContext dbContext, IMapper mapper, UserManager<User> userManager, IPasswordHasher<User> passwordHasher)
+        public UserRepository(AppDbContext dbContext, UserManager<User> userManager, IPasswordHasher<User> passwordHasher)
         {
             _dbContext = dbContext;
-            _mapper = mapper;
             _userManager = userManager;
             _passwordHasher = passwordHasher;
         }
 
-        public async Task<UserDTO> GetUser(string userId)
+        public async Task<User?> GetUser(string userId)
         {
             var user = await _dbContext.Users.FindAsync(userId);
-            var userDTO = _mapper.Map<UserDTO>(user);
-            return userDTO;
+            return user;
         }
 
-        public async Task<IEnumerable<UserDTO>> GetUsers()
+        public async Task<IEnumerable<User>> GetUsers()
         {
             var users = await _dbContext.Users.OrderByDescending(u => u.CreationDate).ToListAsync();
-            var userDTOs = _mapper.Map<IEnumerable<UserDTO>>(users);
-            return userDTOs;
+            return users;
         }
 
-        public async Task<IEnumerable<UserDTO>> GetOnlyUsers()
+        public async Task<IEnumerable<User>> GetOnlyUsers()
         {
             var onlyUsers = await _userManager.GetUsersInRoleAsync("User");
             var sortedUsers = onlyUsers.OrderByDescending(u => u.CreationDate);
-            var onlyUserDTOs = _mapper.Map<IEnumerable<UserDTO>>(sortedUsers);
-            return onlyUserDTOs;
+            return sortedUsers;
         }
 
-        public async Task<IEnumerable<UserDTO>> GetOnlyAdmins()
+        public async Task<IEnumerable<User>> GetOnlyAdmins()
         {
             var onlyAdmins = await _userManager.GetUsersInRoleAsync("Admin");
             var sortedUsers = onlyAdmins.OrderByDescending(u => u.CreationDate);
-            var onlyAdminDTOs = _mapper.Map<IEnumerable<UserDTO>>(sortedUsers);
-            return onlyAdminDTOs;
+            return sortedUsers;
         }
 
-        public async Task<string> GetUserRole(string userId)
+        public async Task<string?> GetUserRole(string userId)
         {
             return await (from ur in _dbContext.UserRoles
                                   join r in _dbContext.Roles on ur.RoleId equals r.Id
@@ -67,29 +59,21 @@ namespace ChatLab.CoreService.Repositories.Classes
             return await _dbContext.Users.CountAsync();
         }
 
-        public async Task UpdateUser(string userId, UserUpdateDTO dto)
+        public async Task UpdateUser(User user)
         {
-            var user = await _dbContext.Users.FindAsync(userId);
             if (user != null)
             {
-                _mapper.Map(dto, user);
                 _dbContext.Entry(user).State = EntityState.Modified;
                 await _dbContext.SaveChangesAsync();
             }
         }
 
-        public async Task ResetUserPassword(string userId, UserResetPasswordDTO dto)
+        public async Task ResetUserPassword(User user, string newPassword)
         {
-            if (!dto.PasswordHash.Equals(dto.ConfirmPasswordHash))
-                throw new ArgumentException($"Confirmed password is different.");
-
-            var user = await _dbContext.Users.FindAsync(userId);
             if (user != null)
             {
-                _mapper.Map(dto, user);
-
-                if (!string.IsNullOrEmpty(dto.PasswordHash))
-                    user.PasswordHash = _passwordHasher.HashPassword(user, dto.PasswordHash);
+                if (!string.IsNullOrEmpty(newPassword))
+                    user.PasswordHash = _passwordHasher.HashPassword(user, newPassword);
 
                 _dbContext.Entry(user).State = EntityState.Modified;
                 await _dbContext.SaveChangesAsync();
@@ -119,6 +103,11 @@ namespace ChatLab.CoreService.Repositories.Classes
                 }
             }
             _dbContext.Chats.RemoveRange(chats);
+
+            var userFollowers = await _dbContext.UserFollowers
+                .Where(f => f.FollowerId == userId || f.FollowedId == userId)
+                .ToListAsync();
+            _dbContext.UserFollowers.RemoveRange(userFollowers);
 
             var unknownUser = await _dbContext.Users
                .Where(u => u.Email == "unknown@unknown.com")
@@ -161,6 +150,15 @@ namespace ChatLab.CoreService.Repositories.Classes
                 .OrderByDescending(c => c.LastMessageTimestamp)
                 .Select(c => c.Chat)
                 .ToList();
+        }
+
+        public async Task<IEnumerable<UserFollow>> GetUserFollowersForUser(string userId)
+        {
+            return await _dbContext.UserFollowers
+                .Include(f => f.Follower)
+                .Include(f => f.Followed)
+                .Where(f => f.FollowerId == userId || f.FollowedId == userId)
+                .ToListAsync();
         }
 
         public async Task<MemoryStream> ExportUsersToCsv()
